Experiment with Apple's VideoToolbox API on MacOS.
Simple smoke test of the decode call sequence.

The MacOS 10.8 release notes call this the 64-bit replacement
for the earlier quicktime decoder API, but the only documentation
I can find is in the framework headers, which do have brief function
and parameter descriptions.

I hear a beta version of this framework is present on MacOS 10.7 as well,
but its entry points are on my 10.8.5 system are marked as new in 10.8.