Fix ftyp brand enumeration.
The size field of a box includes the size and the type fields, so
we need to subtract 8+8 bytes to get the length of the brand table,
not just the 8 bytes of the other fixed fields.

Also print '' for empty brands (only happens with the bug above).

And stop dumping the mdat chunk contents.